requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-Cors
Flask-Session
PyJWT
marshmallow
bcrypt
psutil